Moving qubit allocation to qirlib
