+++ qirlib/src/codegen/mod.rs
@@ -7,6 +7,7 @@ use std::path::Path;
 pub mod basicvalues;
 pub mod calls;
 pub mod qis;
+pub mod qubits;
 pub mod rt;
 pub mod types;
